backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.4.2
pydantic-settings==2.1.0
email-validator==2.1.0
python-multipart==0.0.6